feat: extract `Keypoints` into a registry block
First Tier 2 block of the pmndrs UI distribution, and the first thing to
resolve a dependency across registries: `bg-surface-dim` has no shadcn
equivalent, so the item declares `pmndrs/design-system/md3#v0.3.0` and the
install pulls Tier 1 with it. That path is what the whole architecture rests
on, and nothing else exercises it.

The component moves to `registry/keypoints/`, at the root where shadcn looks
for registry items, and this app imports it back through a `@/registry/*`
alias — so what ships is the file the site renders, not a copy of it.

Three changes were needed to make it standalone:

- it pulled `ul` and `li` from this app's MDX component map. Both are trivial
  wrappers and no second block needs them yet, so they are inlined; the rule
  is to promote them to their own item at n=2.
- `className` now merges after the defaults rather than before, so a consumer
  can override them, and `title` is optional in the type as it already was in
  practice.
- `border-outline-variant` becomes `border-border`. The two are 1:1, and the
  stock token is the one a consumer who skipped Tier 1 still has.
  `bg-surface-dim` stays: it is the deliberate reach, and the reason for the
  dependency.

`registry.json` is validated by `pnpm check-registry`. That validator never
parses TSX, so it cannot see a missing token — an install into a scratch app
outside this repo remains the only check that would.
