+++ package.json
@@ -39,6 +39,7 @@
     "prettier-plugin-tailwindcss": "^0.7.2",
     "rehype-parse": "^9.0.1",
     "rehype-stringify": "^10.0.1",
+    "shadcn": "^4.18.0",
     "storybook": "^10.2.0",
     "tailwindcss": "^4.1.18",
     "tailwindcss-animate": "^1.0.7",
@@ -89,8 +90,9 @@
     "dev": "NODE_OPTIONS='--inspect' next",
     "start": "next start",
     "build": "./next-build.sh",
-    "format": "prettier -w src/",
-    "lint": "eslint src/**/*.{ts,tsx} && prettier . --check",
+    "format": "prettier -w src/ registry/",
+    "lint": "eslint src/**/*.{ts,tsx} registry/ && prettier . --check",
+    "check-registry": "shadcn registry validate ./registry.json",
     "prepare": "husky && (test -f scripts/copy-fonts.sh && sh scripts/copy-fonts.sh || true)",
     "release": "changeset publish",
     "changeset": "changeset",
