--- src/components/mdx/Keypoints/index.ts
@@ -1,1 +1,0 @@
-export * from './Keypoints'
