+++ src/utils/compileMdxContent.tsx
@@ -33,7 +33,7 @@ import { Img } from '@/components/mdx/Img'
 import { rehypeImg } from '@/components/mdx/Img/rehypeImg'
 import { Intro } from '@/components/mdx/Intro'
 import { rehypeLink } from '@/components/mdx/Link/rehypeLink'
-import { Keypoints, KeypointsItem } from '@/components/mdx/Keypoints'
+import { Keypoints, KeypointsItem } from '@/registry/keypoints/keypoints'
 import { Mermaid } from '@/components/mdx/Mermaid'
 import { rehypeMermaid } from '@/components/mdx/Mermaid/rehypeMermaid'
 import { Backers, Contributors } from '@/components/mdx/People'
