+++ vitest.config.ts
@@ -7,6 +7,7 @@ export default defineConfig({
       // Mirrors the tsconfig `paths`. Must come before the `@` prefix alias, which
       // would otherwise resolve this to src/package.json and fail to import ./route.
       '@/package.json': path.resolve(__dirname, './package.json'),
+      '@/registry': path.resolve(__dirname, './registry'),
       '@': path.resolve(__dirname, './src'),
     },
   },
